requirements.txt
sarvamai
openai-whisper
python-dotenv
sounddevice
scipy
numpy
pvporcupine
webrtcvad-wheels
websockets
pyautogui
pygetwindow
pyttsx3
requests
beautifulsoup4
google-generativeai